REFACTOR: 파일 관리 Repository 생성을 위한 인터페이스 추출
