+++ src/test/java/com/programmers/springbootbasic/domain/voucher/service/VoucherServiceTest.java
@@ -2,6 +2,7 @@ package com.programmers.springbootbasic.domain.voucher.service;
 
 import com.programmers.springbootbasic.domain.voucher.dto.VoucherRequestDto;
 import com.programmers.springbootbasic.domain.voucher.entity.Voucher;
+import com.programmers.springbootbasic.domain.voucher.repository.MemoryVoucherRepository;
 import com.programmers.springbootbasic.domain.voucher.repository.VoucherRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -19,7 +20,7 @@ class VoucherServiceTest {
 
     @BeforeEach
     void init() {
-        voucherRepository = new VoucherRepository();
+        voucherRepository = new MemoryVoucherRepository();
         voucherService = new VoucherService(voucherRepository);
     }
